requirements.txt
pandas
requests
python-dotenv
pyarrow
duckdb
flatten_json
streamlit
plotly